stepik-boomstream-v2/requirements.txt
flask
sqlalchemy
psycopg2-binary
requests
gunicorn
werkzeug
python-dotenv
python-telegram-bot==20.8
httpx~=0.26.0
